fixes to ssh key generation
